Split out BuildingType into BuildingType.{h,cpp}
