+++ Empire/ProductionQueue.cpp
@@ -1,7 +1,7 @@
 #include "ProductionQueue.h"
 
 #include "Empire.h"
-#include "../universe/Building.h"
+#include "../universe/BuildingType.h"
 #include "../universe/Condition.h"
 #include "../universe/ShipPart.h"
 #include "../universe/ShipPartHull.h"
