+++ UI/BuildingsPanel.cpp
@@ -8,6 +8,7 @@
 #include "../util/OptionsDB.h"
 #include "../util/Order.h"
 #include "../universe/Building.h"
+#include "../universe/BuildingType.h"
 #include "../universe/Effect.h"
 #include "../universe/Planet.h"
 #include "../universe/Enums.h"
