+++ universe/Building.h
@@ -3,24 +3,8 @@
 
 #include "UniverseObject.h"
 #include "ObjectMap.h"
-#include "CommonParams.h"
 #include "../util/Export.h"
-#include "../util/Pending.h"
 
-#include <boost/algorithm/string/case_conv.hpp>
-
-class BuildingType;
-
-namespace Effect {
-    class EffectsGroup;
-}
-namespace Condition {
-    struct Condition;
-}
-namespace ValueRef {
-    template <typename T>
-    struct ValueRef;
-}
 
 /** A Building UniverseObject type. */
 class FO_COMMON_API Building : public UniverseObject {
@@ -82,151 +66,5 @@ private:
     void serialize(Archive& ar, const unsigned int version);
 };
 
-/** A specification for a building of a certain type.  Each building type must
-  * have a \a unique name string, by which it can be looked up using
-  * GetBuildingType(...). */
-class FO_COMMON_API BuildingType {
-public:
-    /** \name Structors */ //@{
-    BuildingType(const std::string& name,
-                 const std::string& description,
-                 CommonParams& common_params,
-                 CaptureResult capture_result,
-                 const std::string& icon);
-
-    ~BuildingType();
-    //@}
-
-    /** \name Accessors */ //@{
-    const std::string&              Name() const            { return m_name; }              ///< returns the unique name for this type of building
-    const std::string&              Description() const     { return m_description; }       ///< returns a text description of this type of building
-    std::string                     Dump(unsigned short ntabs = 0) const;                   ///< returns a data file format representation of this object
-
-    bool                            ProductionCostTimeLocationInvariant() const;            ///< returns true if the production cost and time are invariant (does not depend on) the location
-    float                           ProductionCost(int empire_id, int location_id) const;   ///< returns the number of production points required to build this building at this location by this empire
-    float                           PerTurnCost(int empire_id, int location_id) const;      ///< returns the maximum number of production points per turn that can be spend on this building
-    int                             ProductionTime(int empire_id, int location_id) const;   ///< returns the number of turns required to build this building at this location by this empire
-
-    const ValueRef::ValueRef<double>*   Cost() const    { return m_production_cost.get(); } ///< returns the ValueRef that determines ProductionCost()
-    const ValueRef::ValueRef<int>*      Time() const    { return m_production_time.get(); } ///< returns the ValueRef that determines ProductionTime()
-
-    bool                            Producible() const  { return m_producible; }            ///< returns whether this building type is producible by players and appears on the production screen
-
-    const ConsumptionMap<MeterType>&    ProductionMeterConsumption() const      { return m_production_meter_consumption; }
-    const ConsumptionMap<std::string>&  ProductionSpecialConsumption() const    { return m_production_special_consumption; }
-
-    const std::set<std::string>&    Tags() const        { return m_tags; }
-    const Condition::Condition* Location() const        { return m_location.get(); }    ///< returns the condition that determines the locations where this building can be produced
-
-    /** Returns a condition that can be used by the UI to further filter (beyond
-      * the Location() requirement) where this building will be presented for
-      * enqueuing onto the production queue, to avoid clutter in the
-      * BuildDesignatorWnd. Example usage: Buildings that are already enqueued
-      * at a production location are hidden so they don't appear in the list of
-      * available items that can be enqueued/produced (again) at that location. */
-    const Condition::Condition* EnqueueLocation() const { return m_enqueue_location.get(); }
-
-    /** Returns the EffectsGroups that encapsulate the effects that buildings ofi
-        this type have when operational. */
-    const std::vector<std::shared_ptr<Effect::EffectsGroup>>& Effects() const
-    { return m_effects; }
-
-    const std::string&              Icon() const            { return m_icon; }              ///< returns the name of the grapic file for this building type
-
-    bool ProductionLocation(int empire_id, int location_id) const;  ///< returns true iff the empire with ID empire_id can produce this building at the location with location_id
-    bool EnqueueLocation(int empire_id, int location_id) const;     ///< returns true iff the empire with ID empire_id meets the requirements of the EnqueueLocation() UI filter method for this building at the location with location_id
-
-    /** returns CaptureResult for empire with ID \a to_empire_id capturing from
-      * empire with IDs \a from_empire_id the planet (or other UniverseObject)
-      * with id \a location_id on which this type of Building is located (if 
-      * \a as_production_item is false) or which is the location of a Production
-      * Queue BuildItem for a building of this type (otherwise) */
-    CaptureResult                   GetCaptureResult(int from_empire_id, int to_empire_id,
-                                                     int location_id, bool as_production_item) const
-    { return m_capture_result; }
-
-    /** Returns a number, calculated from the contained data, which should be
-      * different for different contained data, and must be the same for
-      * the same contained data, and must be the same on different platforms
-      * and executions of the program and the function. Useful to verify that
-      * the parsed content is consistent without sending it all between
-      * clients and server. */
-    unsigned int GetCheckSum() const;
-    //@}
-
-private:
-    void Init();
-
-    std::string                                         m_name;
-    std::string                                         m_description;
-    std::unique_ptr<ValueRef::ValueRef<double>>         m_production_cost;
-    std::unique_ptr<ValueRef::ValueRef<int>>            m_production_time;
-    bool                                                m_producible = true;
-    CaptureResult                                       m_capture_result;
-    std::set<std::string>                               m_tags;
-    ConsumptionMap<MeterType>                           m_production_meter_consumption;
-    ConsumptionMap<std::string>                         m_production_special_consumption;
-    std::unique_ptr<Condition::Condition>               m_location;
-    std::unique_ptr<Condition::Condition>               m_enqueue_location;
-    std::vector<std::shared_ptr<Effect::EffectsGroup>>  m_effects;
-    std::string                                         m_icon;
-};
-
-/** Holds all FreeOrion building types.  Types may be looked up by name. */
-class BuildingTypeManager {
-public:
-    using BuildingTypeMap = std::map<std::string, std::unique_ptr<BuildingType>>;
-    using iterator = BuildingTypeMap::const_iterator;
-
-    /** \name Accessors */ //@{
-    /** returns the building type with the name \a name; you should use the
-      * free function GetBuildingType(...) instead, mainly to save some typing. */
-    const BuildingType*         GetBuildingType(const std::string& name) const;
-
-    /** iterator to the first building type */
-    FO_COMMON_API iterator                    begin() const;
-
-    /** iterator to the last + 1th building type */
-    FO_COMMON_API iterator                    end() const;
-
-    /** returns the instance of this singleton class; you should use the free
-      * function GetBuildingTypeManager() instead */
-    static BuildingTypeManager& GetBuildingTypeManager();
-
-    /** Returns a number, calculated from the contained data, which should be
-      * different for different contained data, and must be the same for
-      * the same contained data, and must be the same on different platforms
-      * and executions of the program and the function. Useful to verify that
-      * the parsed content is consistent without sending it all between
-      * clients and server. */
-    unsigned int                GetCheckSum() const;
-    //@}
-
-    /** Sets building types to the future value of \p pending_building_types. */
-    FO_COMMON_API void SetBuildingTypes(Pending::Pending<BuildingTypeMap>&& pending_building_types);
-
-private:
-    BuildingTypeManager();
-
-    /** Assigns any m_pending_building_types to m_bulding_types. */
-    void CheckPendingBuildingTypes() const;
-
-    /** Future building type being parsed by parser.  mutable so that it can
-        be assigned to m_building_types when completed.*/
-    mutable boost::optional<Pending::Pending<BuildingTypeMap>> m_pending_building_types = boost::none;
-
-    /** Set of building types.  mutable so that when the parse complete it can
-        be updated. */
-    mutable BuildingTypeMap m_building_types;
-
-    static BuildingTypeManager* s_instance;
-};
-
-/** returns the singleton building type manager */
-FO_COMMON_API BuildingTypeManager& GetBuildingTypeManager();
-
-/** Returns the BuildingType specification object for a building of
-  * type \a name.  If no such BuildingType exists, 0 is returned instead. */
-FO_COMMON_API const BuildingType* GetBuildingType(const std::string& name);
 
 #endif // _Building_h_
